data_export_tool/requirements.txt
cx_Freeze==7.2.7
ethiopian_date==1.0
mysql_connector_repackaged==0.3.1
